Some progress, finally.
A bit of futzing in Makefile.inc
Nascent https-everywhere port
Version bump for tor-browser.  Also vetted all of the dependencies
introduced by Makefile.inc etc. to make sure we are in sync with
www/firefox-esr.  Checked patches as well.

TBB 4.0.8 is on deck now, the challenge is to get all the rest of
the pieces done before they release another version.