Major work done on finishing a real release for 4.5.3
Includes a new script, start-tor-browser, which attempts to be a
bash-less replacement for the one that comes with the bundle, and
which also deals with the remaining bundle-related issues explicitly.

Change tor.rc to torrc everywhere, since that was a dumb thing to do.

Tried updating PLISTs by hand this time, maybe that's a better
approch.