Fix dependencies after Daniel Jakots <vigdis@chown.me> pointed out the issue
The run dependences were wrong everywhere, basically.  I should follow
the model of mail/enigmail with the extensions: they don't depend on
tor-browser, they only depend on what they need.  It doesn't make any
sense to install them w/o tor-browser but that's no need for a dep.
The tor-browser package instead run-depends on all of the extensions.
That way we don't need the meta package (we never needed it, it was lame).
Instead you just install tor-browser and all is well.