De-cargo-cultifying of the Makefiles and start the move to TBB 5.0.6
Starting with this commit www/tbb/tor-browser/Makefile depends on
www/mozilla/mozilla.port.mk, which must be patched to support us.

The goal is to get this into the tree so let's get it into shape for
that and stop screwing around.  For the nonce I'm checking in my
hacked copy of www/mozilla/mozilla.port.mk but this will disappear
once we make it acceptable to ports@... just need a place to stash
my work for now.