Completely reverse rdeps after comparing to other similar situations.
The tor-browser package now rdeps on everything needed to complete it:
tor itself and the four browser extension packages in www/tbb.  This
means just pkg_add tor-browser results in a working install.  It also
means that pkg_delete tor-browser will leave the deps installed, but
that's no different than many situations and is what pkg_delete -a is
for.