Yet more tightening up
Patch tor-launcher to not send e.g. --geoip to its instance of tor and
instead set GeoIPData and GeoIPv6Data in torrc-defaults so that we don't
need to copy these files - just use them in /usr/local/share/tor
This also means tor-browser.cfg doesn't need to deal with this.
While we're at it get rid of tor-launcher/files/torrc since we don't
need it.
The start-tor-browser script is officially dead.